+++ src/gospatial/app/db_test.go
@@ -8,7 +8,7 @@ import (
 // go test -bench=.
 // go test -bench=. -test.benchmem
 
-// Test NewLayer 
+// Test NewLayer
 // Benchmark InsertFeature
 // Test InsertFeature
 
@@ -22,9 +22,10 @@ const (
 // Benchmark Database.InsertCustomer
 /*=======================================*/
 func BenchmarkDbInsertCustomer(b *testing.B) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	test_customer := Customer{Apikey: test_customer_apikey}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -36,9 +37,10 @@ func BenchmarkDbInsertCustomer(b *testing.B) {
 // Benchmark Database.getCustomer
 /*=======================================*/
 func BenchmarkDbGetCustomerWithCache(b *testing.B) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	test_customer := Customer{Apikey: test_customer_apikey}
 	test_db.InsertCustomer(test_customer)
 	b.ResetTimer()
@@ -48,9 +50,10 @@ func BenchmarkDbGetCustomerWithCache(b *testing.B) {
 }
 
 func BenchmarkDbGetCustomerWithOutCache(b *testing.B) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	test_customer := Customer{Apikey: test_customer_apikey}
 	test_db.InsertCustomer(test_customer)
 	b.ResetTimer()
@@ -65,9 +68,10 @@ func BenchmarkDbGetCustomerWithOutCache(b *testing.B) {
 // Unittest Database.InsertCustomer
 /*=======================================*/
 func TestDbCustomers(t *testing.T) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	test_customer := Customer{Apikey: test_customer_apikey}
 	err := test_db.InsertCustomer(test_customer)
 	if err != nil {
@@ -86,9 +90,10 @@ func TestDbCustomers(t *testing.T) {
 // Benchmark Database.NewLayer
 /*=======================================*/
 func BenchmarkDbNewLayer(b *testing.B) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		test_db.NewLayer()
@@ -99,9 +104,10 @@ func BenchmarkDbNewLayer(b *testing.B) {
 // Benchmark Database.InsertLayer
 /*=======================================*/
 func BenchmarkDbInsertLayer(b *testing.B) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	data := []byte(`{"crs":{"properties":{"name":"urn:ogc:def:crs:OGC:1.3:CRS84"},"type":"name"},"features":[{"geometry":{"coordinates":[[[-76.64062,50.73645513701065],[-76.64062,65.65827451982659],[-38.67187,65.65827451982659],[-38.67187,50.73645513701065],[-76.64062,50.73645513701065]]],"type":"Polygon"},"properties":{"FID":0},"type":"Feature"},{"geometry":{"coordinates":[[[-87.97851562499999,58.995311187950925],[-87.97851562499999,60.500525410511294],[-84.63867187499997,60.500525410511294],[-84.63867187499997,58.995311187950925],[-87.97851562499999,58.995311187950925]]],"type":"Polygon"},"properties":{"FID":1},"type":"Feature"}],"type":"FeatureCollection"}`)
 	geojs, err := geojson.UnmarshalFeatureCollection(data)
 	if err != nil {
@@ -117,10 +123,10 @@ func BenchmarkDbInsertLayer(b *testing.B) {
 // Benchmark Database.GetLayer
 /*=======================================*/
 func BenchmarkDbGetLayerWithCache(b *testing.B) {
-	TestMode()
-	// StandardMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	data := []byte(`{"crs":{"properties":{"name":"urn:ogc:def:crs:OGC:1.3:CRS84"},"type":"name"},"features":[{"geometry":{"coordinates":[[[-76.64062,50.73645513701065],[-76.64062,65.65827451982659],[-38.67187,65.65827451982659],[-38.67187,50.73645513701065],[-76.64062,50.73645513701065]]],"type":"Polygon"},"properties":{"FID":0},"type":"Feature"},{"geometry":{"coordinates":[[[-87.97851562499999,58.995311187950925],[-87.97851562499999,60.500525410511294],[-84.63867187499997,60.500525410511294],[-84.63867187499997,58.995311187950925],[-87.97851562499999,58.995311187950925]]],"type":"Polygon"},"properties":{"FID":1},"type":"Feature"}],"type":"FeatureCollection"}`)
 	geojs, err := geojson.UnmarshalFeatureCollection(data)
 	if err != nil {
@@ -137,10 +143,10 @@ func BenchmarkDbGetLayerWithCache(b *testing.B) {
 // Benchmark Database.GetLayer
 /*=======================================*/
 func BenchmarkDbGetLayerWithoutCache(b *testing.B) {
-	TestMode()
-	// StandardMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	data := []byte(`{"crs":{"properties":{"name":"urn:ogc:def:crs:OGC:1.3:CRS84"},"type":"name"},"features":[{"geometry":{"coordinates":[[[-76.64062,50.73645513701065],[-76.64062,65.65827451982659],[-38.67187,65.65827451982659],[-38.67187,50.73645513701065],[-76.64062,50.73645513701065]]],"type":"Polygon"},"properties":{"FID":0},"type":"Feature"},{"geometry":{"coordinates":[[[-87.97851562499999,58.995311187950925],[-87.97851562499999,60.500525410511294],[-84.63867187499997,60.500525410511294],[-84.63867187499997,58.995311187950925],[-87.97851562499999,58.995311187950925]]],"type":"Polygon"},"properties":{"FID":1},"type":"Feature"}],"type":"FeatureCollection"}`)
 	geojs, err := geojson.UnmarshalFeatureCollection(data)
 	if err != nil {
@@ -159,9 +165,10 @@ func BenchmarkDbGetLayerWithoutCache(b *testing.B) {
 // Unittest: Database.InsertLayer
 /*=======================================*/
 func TestDbLayers(t *testing.T) {
-	TestMode()
+	test_logger_init()
 	test_db := Database{File: test_db_file}
 	test_db.Init()
+	test_db.TestLogger()
 	data := []byte(`{"crs":{"properties":{"name":"urn:ogc:def:crs:OGC:1.3:CRS84"},"type":"name"},"features":[{"geometry":{"coordinates":[[[-76.64062,50.73645513701065],[-76.64062,65.65827451982659],[-38.67187,65.65827451982659],[-38.67187,50.73645513701065],[-76.64062,50.73645513701065]]],"type":"Polygon"},"properties":{"FID":0},"type":"Feature"},{"geometry":{"coordinates":[[[-87.97851562499999,58.995311187950925],[-87.97851562499999,60.500525410511294],[-84.63867187499997,60.500525410511294],[-84.63867187499997,58.995311187950925],[-87.97851562499999,58.995311187950925]]],"type":"Polygon"},"properties":{"FID":1},"type":"Feature"}],"type":"FeatureCollection"}`)
 	geojs, err := geojson.UnmarshalFeatureCollection(data)
 	if err != nil {